// app-next-daakiya-release/src/application/constants/enums/loading.state.ts
export enum LoadingState {
   DEFAULT = 'DEFAULT',
   LIST_CAMPAIGN = 'LIST_CAMPAIGN',
   LIST_COHORT = 'LIST_COHORT',
   CREATE_COHORT = 'CREATE_COHORT',
   UPDATE = 'UPDATE',
   GET_CAMPAIGN = 'GET_CAMPAIGN',
   GET_CAMPAIGN_STATS = 'GET_CAMPAIGN_STATS',
   CREATE_CAMPAIGN = 'CREATE_CAMPAIGN',
   CREATE_PUSH_NOTIFICATION = 'CREATE_PUSH_NOTIFICATION',
   OVERVIEW = 'OVERVIEW',
}